requirements.txt
beautifulsoup4==4.12.2
python-pptx==0.6.21
requests==2.31.0
tinycss2==1.2.1
Pillow==10.0.0